requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.1
Django==4.0.7
djangorestframework==3.13.1
idna==3.3
numpy==1.23.3
pandas==1.4.4
pyproj==3.3.1
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
scipy==1.9.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.12
